refactor store
